+++ Client/TCPLib.Client/DPDispatcher/DPDispatcherBuilder.cs
@@ -8,22 +8,26 @@ using TCPLib.Net.DPDispatcher;
 
 namespace TCPLib.Client.DPDispatcher
 {
-    
+    /// <summary>
+    /// Data Package Dispatcher
+    /// This class is designed to listen for packets and may be more convenient in some cases.
+    /// For more information, visit the <a href="https://github.com/nyarkus/TCPLib/blob/master/documentation/DataPackageDispatcher.md">Documentation</a>.
+    /// </summary>
     public class DPDispatcherBuilder
     {
         private readonly TCPLib.Client.Net.Server _client;
-        private List<DPHandlerRegistry> _handlers = new List<DPHandlerRegistry>();
+        private List<DPHandler> _handlers = new List<DPHandler>();
         public bool UseDecryption { get; set; } = true;
         public bool ThrowIfNotHandled { get; set; } = true;
-
-        public DPDispatcherBuilder(TCPLib.Client.Net.Server client, params DPHandlerRegistry[] handlers)
+        /// <param name="client">The client must be connected to the server</param>
+        public DPDispatcherBuilder(TCPLib.Client.Net.Server client, params DPHandler[] handlers)
         {
             _client = client;
-            _handlers = new List<DPHandlerRegistry>(handlers);
+            _handlers = new List<DPHandler>(handlers);
         }
-        public DPDispatcherBuilder AddDataPackageHandlerRegistry(DPHandlerRegistry handler)
+        public DPDispatcherBuilder AddDataPackageHandlerRegistry(DPHandler handler)
         {
-            var handlers = new List<DPHandlerRegistry>(_handlers)
+            var handlers = new List<DPHandler>(_handlers)
             {
                 handler
             };
@@ -38,7 +42,7 @@ namespace TCPLib.Client.DPDispatcher
             return new DPDispatcher(_client, handlers, UseDecryption, ThrowIfNotHandled);
         }
 
-        private static void CheckDuplications(DPHandlerRegistry[] handlers)
+        private static void CheckDuplications(DPHandler[] handlers)
         {
             var uniqueHandlers = new HashSet<DPFilter>();
