+++ Server/TCPLib.Server/DPDispatcher/DPDispatcher.cs
@@ -7,10 +7,15 @@ using System.Threading.Tasks;
 
 namespace TCPLib.Server.DPDispatcher
 {
+    /// <summary>
+    /// Data Package Dispatcher
+    /// This class is designed to listen for packets and may be more convenient in some cases.
+    /// For more information, visit the <a href="https://github.com/nyarkus/TCPLib/blob/master/documentation/DataPackageDispatcher.md">Documentation</a>.
+    /// </summary>
     public class DPDispatcher : IDisposable
     {
         private readonly TCPLib.Server.Net.Client _client;
-        private readonly DPHandlerRegistry[] _handlers;
+        private readonly DPHandler[] _handlers;
 
         public bool UseDecryption { get; set; }
         public bool ThrowIfNotHandled { get; set; }
@@ -52,7 +57,7 @@ namespace TCPLib.Server.DPDispatcher
             _cancellationTokenSource.Dispose();
         }
 
-        internal DPDispatcher(TCPLib.Server.Net.Client client, DPHandlerRegistry[] handlers, bool UseDecryption, bool ThrowIfNotHandled)
+        internal DPDispatcher(TCPLib.Server.Net.Client client, DPHandler[] handlers, bool UseDecryption, bool ThrowIfNotHandled)
         {
             _client = client;
             _handlers = handlers;
