--- Server/TCPLib.Server/Exceptions.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace TCPLib.Server
-{
-    public class ServerIsNotRunning : Exception
-    {
-        public ServerIsNotRunning() : base("The server is not running") { }
-    }
-}
