+++ Server/TCPLib.Server/ServerComponents.cs
@@ -10,7 +10,13 @@ namespace TCPLib.Server
     public enum ServerComponents
     {
         None = 0,
+        /// <summary>
+        /// These are commands like a ban, stop, kick etc.
+        /// </summary>
         BaseCommands = 1 << 0,
+        /// <summary>
+        /// If the server receives any packet via UDP, it will send information about itself (Name, description, number of players and their maximum number). 
+        /// </summary>
         UDPStateSender = 1 << 1,
         All = 1 << 2,
     }
